pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "this-day-in-zoom"
version = "1.0.0"
description = "AI-generated Zoom virtual backgrounds based on historical events that happened on this day."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "This Day in Zoom Contributors"}
]
keywords = ["zoom", "virtual-background", "ai", "history", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Conferencing",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "tomli-w>=1.0.0",
    "openai>=1.30.0",
    "anthropic>=0.25.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "respx>=0.21.0",
]

[project.scripts]
tdiz = "tdiz.cli:app"

[project.urls]
Homepage = "https://github.com/your-username/this-day-in-zoom"
Repository = "https://github.com/your-username/this-day-in-zoom"
Issues = "https://github.com/your-username/this-day-in-zoom/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
